Update meow.py
Fixed the connection issues and looping for proxyless users. The config file now auto-updates itself if any settings are missing. Changed failed Hypixel checks to show as 'Unknown' instead of 'Error' to clean up the logs, and fixed the rank colors so they show properly. Code is also cleaned up.